可用版本

+++ CHSNS#1.X/UserWebService/Web/UI/PrivateMaster.cs
@@ -14,7 +14,7 @@ namespace Chsword.Web.UI {
 					}
 				}
 			}
-			byte status = CHUser.Status;
+			int status = CHUser.Status;
 			if (Path.urlFilename == "message") {//MessageË­¶¼ÄÜ¿´
 				return;
 			}
